Revert "Move Pokemon repositories to shared folder in order to avoid code duplication"
This reverts commit 97e6c898d677d7f882c385127ee8a3571b89ce90,
due to bug in serializing data not connected to changes applied during code refactor
