+++ src/main/java/spyra/lukasz/pokerestapi/consume/PokeAbilityRepository.java
@@ -1,11 +1,11 @@
-package spyra.lukasz.pokerestapi.shared.repository;
+package spyra.lukasz.pokerestapi.consume;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import spyra.lukasz.pokerestapi.shared.PokeAbility;
 
 import java.util.Optional;
 
-public interface PokeAbilityRepository extends JpaRepository<PokeAbility, Long> {
+interface PokeAbilityRepository extends JpaRepository<PokeAbility, Long> {
 
     Optional<PokeAbility> findFirstByNameAndHidden(String name, boolean hidden);
