+++ src/main/java/spyra/lukasz/pokerestapi/consume/PokeRepository.java
@@ -1,0 +1,7 @@
+package spyra.lukasz.pokerestapi.consume;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import spyra.lukasz.pokerestapi.shared.Pokemon;
+
+interface PokeRepository extends JpaRepository<Pokemon, Long> {
+}
