+++ src/main/java/spyra/lukasz/pokerestapi/consume/PokeStatRepository.java
@@ -1,11 +1,11 @@
-package spyra.lukasz.pokerestapi.shared.repository;
+package spyra.lukasz.pokerestapi.consume;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import spyra.lukasz.pokerestapi.shared.PokeStat;
 
 import java.util.Optional;
 
-public interface PokeStatRepository extends JpaRepository<PokeStat, Long> {
+interface PokeStatRepository extends JpaRepository<PokeStat, Long> {
 
     Optional<PokeStat> findFirstByNameAndValue(String name, int value);
