+++ src/main/java/spyra/lukasz/pokerestapi/consume/PokeTypeRepository.java
@@ -1,11 +1,11 @@
-package spyra.lukasz.pokerestapi.shared.repository;
+package spyra.lukasz.pokerestapi.consume;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import spyra.lukasz.pokerestapi.shared.PokeType;
 
 import java.util.Optional;
 
-public interface PokeTypeRepository extends JpaRepository<PokeType, Long> {
+interface PokeTypeRepository extends JpaRepository<PokeType, Long> {
 
     Optional<PokeType> findFirstByName(String name);
