+++ src/main/java/spyra/lukasz/pokerestapi/create/PokeCreateRepository.java
@@ -1,0 +1,7 @@
+package spyra.lukasz.pokerestapi.create;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import spyra.lukasz.pokerestapi.shared.Pokemon;
+
+interface PokeCreateRepository extends JpaRepository<Pokemon, Long> {
+}
